requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
deepdiff==4.0.9
Django==2.2.10
django-crispy-forms==1.8.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Flask==1.1.1
Flask-RESTful==0.3.7
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jsonpickle==1.2
MarkupSafe==1.1.1
ordered-set==3.1.1
osquery==3.0.6
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
thrift==0.11.0
urllib3==1.25.7
Werkzeug==0.16.0
